clocks to rxjava2

+++ clocks/src/main/java/com/theopensourcefamily/clocks/Clock.kt
@@ -1,14 +1,14 @@
 package com.theopensourcefamily.clocks
 
-import io.reactivex.rxjava3.core.Observable
-import io.reactivex.rxjava3.core.Scheduler
+import io.reactivex.Observable
+import io.reactivex.Scheduler
 import java.util.concurrent.TimeUnit
 
 class Clock(private val scheduler: Scheduler) {
 
   fun getClockObservable(): Observable<Long> = Observable
     .interval(INIT_DELAY, PERIOD, TimeUnit.MILLISECONDS, scheduler)
-    .filter { it.toLong() % WINDOW == ZERO }
+    .filter { it % WINDOW == ZERO }
 }
 
 const val INIT_DELAY = 10L
